Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flex-aspect-ratio-img-row-013-expected.txt

Pass condition is 3 green 100x100 squares.

Firefox 84a1 passes. Chrome 87 fails them all by making the green rectangles be 200x100.

Transferred size suggestion is the stretched 100px:


Have to subtract the margin from the stretched height to get the transferred size suggestion:


Stretched transferred size suggestion has to obey min-height:



PASS img 1
FAIL img 2 assert_equals:
<img src="support/200x200-green.png" style="margin-bottom: 20px" data-expected-height="100" data-expected-width="100">
width expected 100 but got 120
PASS img 3
